refactor CardEntity
