--- src/main/java/de/gor1lla/cardstalker/domain/MagicCard.java
@@ -1,16 +1,0 @@
-package de.gor1lla.cardstalker.domain;
-
-public class MagicCard extends Card {
-
-    private Character foil;
-
-    @Override
-    public Character getFoil() {
-        return foil;
-    }
-
-    @Override
-    public void setFoil(Character foil) {
-        this.foil = foil;
-    }
-}
